Move flows port interfaces from domain/port to usecase/port - aligns with repo-wide convention
Moved 8 port interfaces (AssemblyPort, CharpadPort, CommandPort, DasmAssemblyPort, ExomizerPort, GoattrackerPort, ImagePort, SpritepadPort) from flows/src/main/kotlin/.../flows/domain/port/ to flows/src/main/kotlin/.../flows/usecase/port/. Updated all 21 consuming files across domain steps, tests, and adapters to use the new package path. Updated arc42 documentation to reflect the new port location and marked technical debt item D4 as resolved.
