+++ doc/arc42/11_risks_and_technical_debt.md
@@ -9,7 +9,7 @@ This is an **honest debt register**. Every item below was verified against the c
 | D1 | **`crunchers` domain missing from the knowledge base** | `doc/kb/domain.md` lists the domains but omits `crunchers` (`crunchers/exomizer` exists in code with `CrunchMemUseCase`/`CrunchRawUseCase`) | Agents/contributors reading only the kb can miss an entire bounded context | Issue: add `crunchers` to `doc/kb/domain.md` (or point kb at this arc42 §5) |
 | D2 | **Inconsistent base package in `spritepad`** ✅ **RESOLVED** | `processors/spritepad` sources previously used package `com.c64lib.rbt.processors.spritepad...` while other domains used `com.github.c64lib.rbt...` — now all `spritepad` source (main, test, `adapters/in/gradle`) and all consumers in `flows/adapters/out/spritepad` use the uniform `com.github.c64lib.rbt.processors.spritepad...` root | Was: breaks package convention. Now: uniform across all domains | Resolved via [PLAN-0004](../../plans/PLAN-0004_rename-spritepad-package-root.md) issue #157 |
 | D3 | **`fllter` typo in a shared package name** | `shared/gradle/.../shared/gradle/fllter/` (contains `BinaryInterleaver.kt`, `Nybbler.kt`) — should be `filter` | Cosmetic but permanent-looking; every import repeats the typo | Issue: rename `fllter` → `filter` (touches importers in `shared:gradle` and infra) |
-| D4 | **Flows ports live under `domain/port`, not `usecase/port`** | Flows port interfaces are at `flows/src/main/kotlin/.../flows/domain/port/` (`AssemblyPort`, `CharpadPort`, `ExomizerPort`, …), whereas other domains place ports under `usecase/port` | Inconsistent port location complicates the "ports sit next to use cases" rule stated in [§8.1](08_crosscutting_concepts.md) | Issue: decide the canonical port location and align (either move flows ports or document the flows-as-orchestrator exception) |
+| D4 | **Flows ports live under `domain/port`, not `usecase/port`** ✅ **RESOLVED** | Flows port interfaces previously lived at `flows/src/main/kotlin/.../flows/domain/port/` (`AssemblyPort`, `CharpadPort`, `ExomizerPort`, …), whereas other domains place ports under `usecase/port` — now all flows ports are at `flows/src/main/kotlin/.../flows/usecase/port/` | Was: inconsistent port location complicates the "ports sit next to use cases" rule. Now: uniform across all domains | Resolved via [PLAN-0005](../../plans/PLAN-0005_move-flows-ports-to-usecase-port.md) issue #159 |
 | D5 | **Single large `afterEvaluate` wiring block** | `RetroAssemblerPlugin.kt` is ~223 lines with all use-case construction and task wiring inline in one `afterEvaluate` | Growing composition root; harder to read as domains are added (a known trade-off of manual DI — see [AD-4](09_architecture_decisions.md)) | Issue: extract per-domain wiring helpers while keeping DI manual and grep-able |
 | D6 | **Stale concept paper** | `doc/concept/*.adoc` covers only ~4 arc42-style chapters and predates `flows`, `crunchers`, and `dasm` | Readers may treat it as current and miss recent subsystems | Issue: mark `doc/concept/` as historical and point it at this arc42 set (this §11 and [README](README.md) already flag it) |
