+++ flows/src/main/kotlin/com/github/c64lib/rbt/flows/usecase/port/AssemblyPort.kt
@@ -22,7 +22,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-package com.github.c64lib.rbt.flows.domain.port
+package com.github.c64lib.rbt.flows.usecase.port
 
 import com.github.c64lib.rbt.flows.domain.config.AssemblyCommand
