+++ plans/EXEC-0005_move-flows-ports-to-usecase-port.md
@@ -1,0 +1,38 @@
+# Execution Log: Move `flows` port interfaces from `domain/port` to `usecase/port`
+
+**Exec ID**: EXEC-0005
+**Plan**: [PLAN-0005](PLAN-0005_move-flows-ports-to-usecase-port.md)
+**Issue**: #159
+**Started**: 2026-07-16
+**Last Updated**: 2026-07-16
+**State**: in progress
+
+## 1. Execution Sessions
+
+### Session 1 — 2026-07-16
+
+- **Scope**: All steps (Phase 1: Steps 1.1–1.5)
+- **Mode**: Per-step
+- **Outcome**: In progress
+
+| Step | Result | Verification | Notes |
+|------|--------|--------------|-------|
+| 1.1 | in_progress | — | Moving 8 port files to usecase/port directory |
+| 1.2 | pending | — | Fixing imports in domain steps and tests |
+| 1.3 | pending | — | Fixing imports and FQN reference in adapters |
+| 1.4 | pending | — | Build verification |
+| 1.5 | pending | — | Documentation updates |
+
+## 2. Deviations from Plan
+
+| # | Step | Deviation | Reason | Impact |
+|---|------|-----------|--------|--------|
+| — | — | None yet | Execution just beginning | — |
+
+## 3. Follow-ups
+
+- None yet
+
+---
+
+**Session Notes:** Starting Phase 1 execution on feature branch `feature/159-move-flows-ports-to-usecase-port`.
