+++ plans/README.md
@@ -9,3 +9,4 @@ Plans are permanent artifacts — do not delete. Terminal plans (`implemented`, 
 | [PLAN-0002](PLAN-0002_arc42-technical-documentation.md) | 2026-07-15 | implemented | arc42 Technical Documentation (Markdown + Mermaid) | #154 | [EXEC-0002](EXEC-0002_arc42-technical-documentation.md) |
 | [PLAN-0003](PLAN-0003_rename-fllter-to-filter.md) | 2026-07-16 | implemented | Rename `fllter` package to `filter` | #158 | [EXEC-0003](EXEC-0003_rename-fllter-to-filter.md) |
 | [PLAN-0004](PLAN-0004_rename-spritepad-package-root.md) | 2026-07-16 | implemented | Rename `processors/spritepad` package root from `com.c64lib` to `com.github.c64lib` | #157 | [EXEC-0004](EXEC-0004_rename-spritepad-package-root.md) |
+| [PLAN-0005](PLAN-0005_move-flows-ports-to-usecase-port.md) | 2026-07-16 | accepted | Move `flows` port interfaces from `domain/port` to `usecase/port` | #159 | [EXEC-0005](EXEC-0005_move-flows-ports-to-usecase-port.md) |
